% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLiNGAM.R
\name{BaseLiNGAM}
\alias{BaseLiNGAM}
\title{LiNGAM class}
\description{
R implementation for LiNGAM models
This class is a base class for LiNGAM.

This code is based on Python implementation:

(The LiNGAM Project: https://sites.google.com/site/sshimizu06/lingam)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{random_state}}{(integer) Random seed}

\item{\code{causal_order}}{(numeric vector) Causal oreder of variables}

\item{\code{adjacency_matrix}}{(numeric matrix) Estimated adjacency matrix}

\item{\code{intercept}}{(numeric vector) Estimated intercept term}

\item{\code{lasso_engine}}{(character) "lars" or "glmnet"}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BaseLiNGAM$new()}}
\item \href{#method-fit}{\code{BaseLiNGAM$fit()}}
\item \href{#method-estimate_adjacency_matrix}{\code{BaseLiNGAM$estimate_adjacency_matrix()}}
\item \href{#method-predict_adaptive_lasso}{\code{BaseLiNGAM$predict_adaptive_lasso()}}
\item \href{#method-clone}{\code{BaseLiNGAM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create lingam object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseLiNGAM$new(random_state = NULL, lasso_engine = "glmnet")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{random_state}}{(integer) Random seed}

\item{\code{lasso_engine}}{(character) "lars" or "glmnet"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
subclasses should implement this method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseLiNGAM$fit(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{(numeric matrix or data.frame) data matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate_adjacency_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-estimate_adjacency_matrix}{}}}
\subsection{Method \code{estimate_adjacency_matrix()}}{
estimate adjacency matrix based on causal order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseLiNGAM$estimate_adjacency_matrix(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{(numeric matrix or data.frame) data matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_adaptive_lasso"></a>}}
\if{latex}{\out{\hypertarget{method-predict_adaptive_lasso}{}}}
\subsection{Method \code{predict_adaptive_lasso()}}{
fit adaptice lasso
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseLiNGAM$predict_adaptive_lasso(X, predictors, target, gamma = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{(numeric matrix or data.frame) data matrix}

\item{\code{predictors}}{(numeric vector) index of explanatory variables}

\item{\code{target}}{(integer) index of target variable}

\item{\code{gamma}}{(numeric) data x will be weighted like x^(gamma) for adaptive lasso}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
coef_ (numeric vector) estimated coefficients
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseLiNGAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
