% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dummy_data.R
\name{gen_dummy_data}
\alias{gen_dummy_data}
\title{generate dummy data}
\usage{
gen_dummy_data(num_samples = 10000, random_state = 1)
}
\arguments{
\item{num_samples}{(integer) number of samples}

\item{random_state}{(integer) random seed}
}
\description{
generates (num_normal + num_anomaly) samples of 6 dimensional data.
}
