% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICALiNGAM.R
\name{ICALiNGAM}
\alias{ICALiNGAM}
\title{ICALiNGAM class}
\description{
R implementation of ICA based LiNGAM algorithm
See reference for details of the algorithm
}
\references{
S. Shimizu, P. O. Hoyer, A. Hyvärinen, and A. J. Kerminen.
A linear non-gaussian acyclic model for causal discovery.
Journal of Machine Learning Research, 7:2003-2030, 2006.
}
\section{Super class}{
\code{\link[rlingam:LiNGAM]{rlingam::LiNGAM}} -> \code{ICALiNGAM}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{max_iter}}{(integer) maximum iterations for fastICA}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ICALiNGAM$new()}}
\item \href{#method-fit}{\code{ICALiNGAM$fit()}}
\item \href{#method-estimate_causal_order}{\code{ICALiNGAM$estimate_causal_order()}}
\item \href{#method-search_causal_order}{\code{ICALiNGAM$search_causal_order()}}
\item \href{#method-clone}{\code{ICALiNGAM$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rlingam" data-topic="LiNGAM" data-id="estimate_adjacency_matrix">}\href{../../rlingam/html/LiNGAM.html#method-estimate_adjacency_matrix}{\code{rlingam::LiNGAM$estimate_adjacency_matrix()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rlingam" data-topic="LiNGAM" data-id="predict_adaptive_lasso">}\href{../../rlingam/html/LiNGAM.html#method-predict_adaptive_lasso}{\code{rlingam::LiNGAM$predict_adaptive_lasso()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create ICALiNGAM object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ICALiNGAM$new(random_state = NULL, lasso_engine = "glmnet", max_iter = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{random_state}}{(integer) Random seed}

\item{\code{lasso_engine}}{(character) "lars" or "glmnet"}

\item{\code{max_iter}}{(integer) maximum iterations of fastICA}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
fit DirectLiNGAM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ICALiNGAM$fit(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{(numeric matrix or data.frame) data matrix to fit
Estimate causal order based on estimated adjacency matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate_causal_order"></a>}}
\if{latex}{\out{\hypertarget{method-estimate_causal_order}{}}}
\subsection{Method \code{estimate_causal_order()}}{
if B is not DAG, set small elements to zero until DAG is obtained
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ICALiNGAM$estimate_causal_order(B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{B}}{(matrix) estimated adjacency matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
integer vector of length(ncol(B))
get causal order of estimated adjacency matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_causal_order"></a>}}
\if{latex}{\out{\hypertarget{method-search_causal_order}{}}}
\subsection{Method \code{search_causal_order()}}{
if given matrix is not DAG, return NULL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ICALiNGAM$search_causal_order(B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{B}}{(matrix) estimated adjacency matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of length(ncol(B)) or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ICALiNGAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
