% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adjacency_mat.R
\name{plot_adjacency_mat}
\alias{plot_adjacency_mat}
\title{Visualize graph based on adjacency matrix}
\usage{
plot_adjacency_mat(adj_mat, node_labels = NULL, round_digit = 2)
}
\arguments{
\item{adj_mat}{(numerical matrix) from: col -> to: row}

\item{node_labels}{(numeric or character vector) node labels}

\item{round_digit}{(integer) maximum digits}
}
\description{
calls DiagrammeR
}
